refactor: readhistory 패키지를 DDD 레이어로 리팩터링
