+++ src/main/java/com/techfork/activity/readhistory/application/command/SaveSearchHistoryCommand.java
@@ -1,0 +1,10 @@
+package com.techfork.activity.readhistory.application.command;
+
+import java.time.LocalDateTime;
+
+public record SaveSearchHistoryCommand(
+        Long userId,
+        String query,
+        LocalDateTime searchedAt
+) {
+}
