+++ src/main/java/com/techfork/activity/readhistory/infrastructure/SearchHistoryRepository.java
@@ -1,12 +1,11 @@
-package com.techfork.activity.readhistory.repository;
+package com.techfork.activity.readhistory.infrastructure;
 
-import com.techfork.activity.readhistory.entity.SearchHistory;
+import com.techfork.activity.readhistory.domain.SearchHistory;
+import java.util.List;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-
-import java.util.List;
 
 public interface SearchHistoryRepository extends JpaRepository<SearchHistory, Long> {
