+++ src/main/java/com/techfork/domain/personalization/service/PersonalizationProfileService.java
@@ -2,8 +2,8 @@ package com.techfork.domain.personalization.service;
 
 import com.techfork.activity.bookmark.domain.Bookmark;
 import com.techfork.activity.bookmark.infrastructure.BookmarkRepository;
-import com.techfork.activity.readhistory.entity.SearchHistory;
-import com.techfork.activity.readhistory.repository.SearchHistoryRepository;
+import com.techfork.activity.readhistory.domain.SearchHistory;
+import com.techfork.activity.readhistory.infrastructure.SearchHistoryRepository;
 import com.techfork.activity.readpost.domain.ReadPost;
 import com.techfork.activity.readpost.infrastructure.ReadPostRepository;
 import com.techfork.domain.post.entity.PostKeyword;
