+++ src/test/java/com/techfork/activity/readhistory/presentation/SearchHistoryRequestTest.java
@@ -1,12 +1,11 @@
-package com.techfork.activity.readhistory.dto;
+package com.techfork.activity.readhistory.presentation;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.json.JsonMapper;
+import java.time.LocalDateTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-
-import java.time.LocalDateTime;
 
 import static org.assertj.core.api.Assertions.assertThat;
